+++ CHANGES.rst
@@ -5,6 +5,8 @@ master (unreleased)
 
 API changes
 -----------
+- Added optional `model` parameter to `Experiment.pred_to_pc` to select only
+certain models and model outputs for the conversion to percent correct.
 
 Enhancements
 ------------
@@ -13,6 +15,9 @@ Enhancements
 signal `x`: If `x` is a matrix, the rows are filtered. If `b` is a matrix,
 each filter is applied to `x`. If both `b` and `x` are matrices with the same
 number of rows, each row of `x` is filtered with the respective row of `b`.
+- The `Experiment` class tries to create the output folder if it does not exist.
+- The speech material name is saved out the output data frame when running a
+speech intelligibility experiment.
 
 Performance
 -----------